try to change word_t to u64

+++ bloomfilter/bloomfilter_runtime_types.hpp
@@ -16,8 +16,8 @@ namespace dtl {
 
 // TODO remove namespace pollution
 using key_t = $u32;
-//using word_t = $u64; // FIXME
-using word_t = $u32;
+using word_t = $u64; // FIXME
+//using word_t = $u32;
 
 template<typename T>
 using hash_fn_0 = dtl::hash::knuth<T>;
